Fixed all device names and plugin names
