+++ source/plugins/mock/light/Dimmer.py
